Eliminating unnecessary speller creation/deletion.
Previously, an internal speller object was created and destroyed for every call to #correct? and #suggestions. Now, an internal speller object is only allocated when the initial FFI::Aspell::Speller object is created and when its configuration changes.

This does make the API more complex since users now need to call #close (or use #open), but the performance gains are significant, yielding up to 300x throughput in testing.