Adds `libaspell` filename to array
* Filename on OSX is `libaspell.dyld`, so the name
has to be added as `libaspell` so it is searched
when looking for the library:

```
Could not open library 'aspell': dlopen(aspell, 5): image not found. (LoadError)
Could not open library 'libaspell.dylib': dlopen(libaspell.dylib, 5): image not found.
Could not open library 'libaspell.so.15': dlopen(libaspell.so.15, 5): image not found.
Could not open library 'libaspell.so.15.dylib': dlopen(libaspell.so.15.dylib, 5): image not found.
Could not open library 'libaspell': dlopen(libaspell, 5): image not found
```